Company_reputation_tracker_v2.1/requirements.txt
streamlit
requests
sqlalchemy
newsapi-python
textblob
openai
python-dotenv
